+++ crates/mofa-foundation/src/gateway/mod.rs
@@ -6,6 +6,4 @@
 
 pub mod rate_limiter;
 
-pub use rate_limiter::{
-    KeyStrategy, RateLimitDecision, RateLimiter, RateLimiterConfig, TokenBucketRateLimiter,
-};
+pub use rate_limiter::TokenBucketRateLimiter;
